# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ServITech API",
        "description": "API documentation for ServITech",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/{version}/support-requests": {
            "get": {
                "tags": [
                    "Support Requests"
                ],
                "summary": "Get all support requests",
                "description": "Display a listing of the resource.",
                "operationId": "588c0f5c9326046c2a9cecb43aac5028",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of support requests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Support requests retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "supportRequests": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CreateSupportRequest"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Support Requests"
                ],
                "summary": "Create a new support request",
                "description": "Store a newly created resource in storage.",
                "operationId": "e111111528f9a0e9b1fcbf7b15564b2f",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSupportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Support request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Support request created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "supportRequest": {
                                                    "$ref": "#/components/schemas/CreateSupportRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/{version}/support-requests/{id}": {
            "get": {
                "tags": [
                    "Support Requests"
                ],
                "summary": "Get a specific support request",
                "description": "Display the specified resource.",
                "operationId": "c775f6f707c95ff486b43e89eec85b01",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the support request",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Support request retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Support request retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "supportRequest": {
                                                    "$ref": "#/components/schemas/CreateSupportRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Support Requests"
                ],
                "summary": "Update a support request",
                "description": "Update the specified resource in storage.",
                "operationId": "660eea955163af59d266360cfebdac98",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the support request",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSupportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Support request updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Support request updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "supportRequest": {
                                                    "$ref": "#/components/schemas/UpdateSupportRequest"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Support Requests"
                ],
                "summary": "Delete a support request",
                "description": "Remove the specified resource from storage.",
                "operationId": "3d032457af2cb6a3dea41eef6937da74",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "API version",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "v1"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the support request",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Support request deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Support request deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateSupportRequest": {
                "required": [
                    "date",
                    "location",
                    "detail"
                ],
                "properties": {
                    "date": {
                        "description": "Class SupportRequestController for managing support requests.",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-29"
                    },
                    "location": {
                        "type": "string",
                        "example": "New York"
                    },
                    "detail": {
                        "type": "string",
                        "example": "The system is not responding."
                    }
                },
                "type": "object"
            },
            "UpdateSupportRequest": {
                "required": [
                    "date",
                    "location",
                    "detail"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-30"
                    },
                    "location": {
                        "type": "string",
                        "example": "Los Angeles"
                    },
                    "detail": {
                        "type": "string",
                        "example": "The issue has been updated."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}